refactor: migrate hooks (use-page, use-page-store) from web/app/ce to web/app/core
