+++ apps/web/ce/components/pages/extra-actions.tsx
@@ -5,7 +5,7 @@
  */
 
 // store
-import type { EPageStoreType } from "@/plane-web/hooks/store";
+import type { EPageStoreType } from "@/hooks/store";
 import type { TPageInstance } from "@/store/pages/base-page";
 
 export type TPageHeaderExtraActionsProps = {
