+++ apps/web/core/components/pages/list/block-item-action.tsx
@@ -15,7 +15,7 @@ import { renderFormattedDate, getFileURL } from "@plane/utils";
 import { useMember } from "@/hooks/store/use-member";
 import { usePageOperations } from "@/hooks/use-page-operations";
 // plane web hooks
-import type { EPageStoreType } from "@/plane-web/hooks/store";
+import type { EPageStoreType } from "@/hooks/store";
 // store
 import type { TPageInstance } from "@/store/pages/base-page";
 // local imports
