+++ apps/web/core/components/pages/modals/delete-page-modal.tsx
@@ -14,8 +14,8 @@ import { getPageName } from "@plane/utils";
 // constants
 // plane web hooks
 import { useAppRouter } from "@/hooks/use-app-router";
-import type { EPageStoreType } from "@/plane-web/hooks/store";
-import { usePageStore } from "@/plane-web/hooks/store";
+import type { EPageStoreType } from "@/hooks/store";
+import { usePageStore } from "@/hooks/store";
 // store
 import type { TPageInstance } from "@/store/pages/base-page";
