+++ apps/web/core/components/pages/navigation-pane/root.tsx
@@ -17,7 +17,7 @@ import { useQueryParams } from "@/hooks/use-query-params";
 // plane web components
 import type { TPageNavigationPaneTab } from "@/plane-web/components/pages/navigation-pane";
 // store
-import type { EPageStoreType } from "@/plane-web/hooks/store";
+import type { EPageStoreType } from "@/hooks/store";
 import type { TPageInstance } from "@/store/pages/base-page";
 // local imports
 import type { TPageRootHandlers } from "../editor/page-root";
