+++ apps/web/core/components/pages/pages-list-main-content.tsx
@@ -19,7 +19,7 @@ import { PageLoader } from "@/components/pages/loaders/page-loader";
 import { useProject } from "@/hooks/store/use-project";
 import { useUserPermissions } from "@/hooks/store/user";
 // plane web hooks
-import { EPageStoreType, usePageStore } from "@/plane-web/hooks/store";
+import { EPageStoreType, usePageStore } from "@/hooks/store";
 
 type Props = {
   children: React.ReactNode;
