+++ apps/web/core/hooks/store/use-page.ts
@@ -8,8 +8,8 @@ import { useContext } from "react";
 // mobx store
 import { StoreContext } from "@/lib/store-context";
 // plane web hooks
-import type { EPageStoreType } from "@/plane-web/hooks/store";
-import { usePageStore } from "@/plane-web/hooks/store";
+import type { EPageStoreType } from "./use-page-store";
+import { usePageStore } from "./use-page-store";
 
 export type TArgs = {
   pageId: string;
