+++ apps/web/core/hooks/use-favorite-item-details.tsx
@@ -19,7 +19,7 @@ import { useModule } from "@/hooks/store/use-module";
 import { useProject } from "@/hooks/store/use-project";
 import { useProjectView } from "@/hooks/store/use-project-view";
 // plane web hooks
-import { EPageStoreType, usePage } from "@/plane-web/hooks/store";
+import { EPageStoreType, usePage } from "@/hooks/store";
 import { useAdditionalFavoriteItemDetails } from "@/plane-web/hooks/use-additional-favorite-item-details";
 
 export const useFavoriteItemDetails = (workspaceSlug: string, favorite: IFavorite) => {
